Store UID/GID into int64
uid_t and gid_t are by definition uint32, but since in the cache DB we
also use -1 as a value to represent non-existent UID/GIDs, sometimes we
need to use signed values as well.

A good compromise is to explicitly use int64 wherever we need signed
values, so that we can safely cast the entire range of valid UIDs
(uint32) into int64 values.

This essentially changes instances of int to int64, so 32-bit platforms
will not attempt to cast uint32 values to int32.

+++ internal/cache/cache.go
@@ -340,8 +340,8 @@ func (c *Cache) Update(ctx context.Context, username, password, homeDirPattern, 
 
 		user = UserRecord{
 			Name:  username,
-			UID:   int(id),
-			GID:   int(id),
+			UID:   int64(id),
+			GID:   int64(id),
 			Home:  home,
 			Shell: shell,
 		}
